Pre-compute the return value of getProject() in order to not have to import "@enkore/spec" inside the project API.
It looks like (because treeshaking is disabled for assets) that importing a single function from "@enkore/spec"
leads to the inclusion of the whole "@enkore/spec" package -- which also includes usage of node modules like node:fs or node:crypto.

In order to keep those out of the end-product, we invoke createEntity beforehand so it's no longer needed inside the public project API ... (for now).