Okay, so we have a bit of an issue here:
The synchronous version of pkg-private.node-fs-find-nearest-file needs to have access to node:fs.

However, 'require' is not defined in ESM files, so we have to load {createRequire} from node:module.

This is an asynchronous operation, which means it needs to be at the top level for the function to remain sync.

But top-level await is not supported by rollup when the output format is IIFE.

So to solve this problem, the implementation of find-nearest-file will use the require() method loaded by the runtime init code.